feat: imporove readability, devide into std and substrate
